+++ extras/create_model.py
@@ -1,277 +1,315 @@
-### First version is going to assume everything is a string
-### also known as string theory:p
+# -*- coding: utf-8 -*-
+"""
+Load first rows from csv, ask some questions
+and generate a models.go to jumpstart
+your lambda_db project for the given csv file
 
-### column with the name "value" or "index" will be used as index
-### else the first column will be set as index, when index is enabled.
-### this can be changed later in the generated model.go file
+models.go contains all the field information
+and functions of rows in your data.
+
+- Repeated option to store repeated
+  values in a map and each individual items
+  only stores uint16 reference to map key.
+
+- BitArray option which is like Repeated
+  value but also creates a map[key]bitmap for all
+  items containing field value. Makes it possible
+  to do fast 'match' lookups.
+
+
+python create_model.py your.csv
+"""
 
 import csv
 import sys
-from filereader import create_reader, supported_fileformats
+
+from re import sub
+from jinja2 import Environment, FileSystemLoader
+
+if '-f' in sys.argv:
+    filename = str(sys.argv[sys.argv.index('-f')+1])
+else:
+    filename = "items.csv"
+
+with open(filename) as f:
+    reader = csv.DictReader(f)
+    row = dict(next(reader))
+
+env = Environment(
+    loader=FileSystemLoader('./templates'),
+)
+
+# keep track of all column names and all original names in csv
+allcolumns = []
+allcolumns_org = []
+repeated = []
+repeated_org = []
+bitarray = []
+bitarray_org = []
+unique = []
+unique = []
+unique_org = []
+ignored = []
+ignored_org = []
+geocolumns = []
+geocolumns_org = []
 
 
-def create_struct(item):
-    start = "type Item struct {\n"
-    # TODO add type
-    lines= [f'{k.capitalize()} string `json:"{k.lower()}"`' for k, v in item.items()]
-    stop = "\n}\n"
-    return start + "\n".join(lines) + stop 
+def gocamelCase(string):
+    """convert string to camelCase
 
-
-def create_columns(item):
-    start = """
-    func (i Item) Columns() []string {
-	return []string{
+    woning_type -> WoningType
     """
-    lines = [f'"{k.lower()}",' for k in item.keys()]
-    stop = """\n}\n}"""
-    return start + "\n".join(lines) + stop
+    string = sub(r"(_|-)+", " ", string).title().replace(" ", "")
+    return string
 
 
-def create_row(item):
-    start = """
-    func (i Item) Row() []string {
-	return []string{
-    """
-    lines = [f"i.{k.capitalize()}," for k in item.keys()]
-    stop = """\n}\n}"""
-    return start + "\n".join(lines) + stop
+# ask some questions about columns.
+index = 0
+for k in row.keys():
 
+    # go camelcase column names
+    kc = gocamelCase(k)
 
-def get_index_column(item):
-    special_columns = ["value", "index"]
-    for column in special_columns:
-        if column in item:
-            return column
+    options = ['r', 'u', 'i', 'g', 'b']
+    while True:
+        # keep asking for valid input
+        q1 = (
+            "(R)epeated value? has less then (2^16=65536) option.",
+            "(B)itarray, repeated column optimized for fast match.",
+            "(U)nique, (G)eo lat/lon point or (I)gnore ? r/b/u/g/i?."
+        )
+        action = input(f"idx:{index} is {k} {q1}")  # noqa
+        if action == '':
+            print(f"pick one from {options}")
+            continue
+        if action not in options:
+            continue
+        break
 
-    # we tried, let's return the first column
-    n = iter(item.keys())
-    return next(n)
+    if action == 'r':
+        repeated.append(kc)
+        repeated_org.append(k)
+    elif action == 'u':
+        unique.append(kc)
+        unique_org.append(k)
+    elif action == 'i':
+        ignored.append(kc)
+        ignored_org.append(k)
+    elif action == 'g':
+        geocolumns.append(kc)
+        geocolumns_org.append(k)
+        unique.append(kc)
+        unique_org.append(k)
+    elif action == 'b':
+        # same as repeated  but with some extra bitarray stuff
+        repeated.append(kc)
+        repeated_org.append(k)
+        bitarray.append(kc)
+        bitarray_org.append(k)
+    else:
+        print('invalid input')
+        sys.exit(-1)
 
+    allcolumns.append(kc)
+    allcolumns_org.append(k)
+    index += 1
 
-def create_getindex(item):
-    index_column = get_index_column(item)
-    start = """
-    func (i Item) GetIndex() string {
-	return """
-    middle = f"i.{index_column.capitalize()}"
-    stop = """\n}"""
-    return start + middle + stop
+# ask for a index column
+while True:
+    # keep asking for valid input
+    index = input(f"which column is idx? 0 - {len(allcolumns) - 1} ")
+    try:
+        index = int(index)
 
+        if allcolumns[index] in ignored:
+            print('Selected an ignored column for index')
+            raise ValueError
 
-def create_filter_contains(column):
-    return (
-            f"func Filter{column.capitalize()}Contains(i *Item, s string) bool"  + "{" + "\n"
-            f"return strings.Contains(i.{column.capitalize()}, s)"
-            "\n" + "}"
+        if -1 < index < len(allcolumns):
+            break
+
+    except ValueError:
+        continue
+
+    print('try again..')
+
+# setup initial data structs for each repeated column
+initRepeatColumns = []
+initColumntemplate = env.get_template('initColumn.template.jinja2')
+
+for c in repeated:
+    initRepeatColumns.append(
+        initColumntemplate.render(
+            columnName=c, bitarraymap=c in bitarray)
     )
 
-def create_filter_startswith(column):
-    return (
-            f"func Filter{column.capitalize()}StartsWith(i *Item, s string) bool"  + "{" + "\n"
-            f"return strings.HasPrefix(i.{column.capitalize()}, s)"
-            "\n" + "}"
-    )
+# create bitarrays with item labels for column values.
+bitArrayStores = []
+bitArrayGetters = []
+bitArrayStoreTemplate = env.get_template('storebitarray.template.jinja2')
+bitArrayGetTemplate = env.get_template('bitarrayGetter.template.jinja2')
+for r in bitarray:
+    bitArrayStores.append(bitArrayStoreTemplate.render(columnName=r))
+    bitArrayGetters.append(bitArrayGetTemplate.render(columnName=r))
 
-def create_filter_match(column):
-    return (
-            f"func Filter{column.capitalize()}Match(i *Item, s string) bool"  + "{" + "\n"
-            f"return i.{column.capitalize()} ==  s"
-            "\n" + "}"
-    )
+# create ItemFull struct fields
+columnsItemIn = []
+jsonColumn = env.get_template('itemFullColumn.jinja2')
+for c1, c2 in zip(allcolumns, allcolumns_org):
+    onerow = jsonColumn.render(c1=c1, c2=c2)
+    columnsItemIn.append(onerow)
 
+# create ItemFull struct fields
+columnsItemOut = []
+jsonColumn = env.get_template('itemFullColumn.jinja2')
+for c1, c2 in zip(allcolumns, allcolumns_org):
 
-def create_getter(column):
-    return (
-            f"func Getters{column.capitalize()}(i *Item) string"  + "{" + "\n"
-            f"return i.{column.capitalize()}"
-            "\n" + "}"
-    )
+    if c1 in ignored:
+        continue
 
+    onerow = jsonColumn.render(c1=c1, c2=c2)
+    columnsItemOut.append(onerow)
 
-def create_reduce(column):
-    return """
-    func reduceCount(items Items) map[string]string {
-	result := make(map[string]string)
-	result["count"] = strconv.Itoa(len(items))
-	return result
-}
-    """
+# create Item struct fields
+columnsItem = []
+for c1, c2 in zip(allcolumns, allcolumns_org):
 
-def create_init_register():
-    return """
-    RegisterFuncMap = make(registerFuncType)
-    RegisterGroupBy = make(registerGroupByFunc)
-    RegisterGetters = make(registerGettersMap)
-    RegisterReduce = make(registerReduce)
+    if c1 in ignored:
+        continue
 
-    """
-
-def create_register_match_func(column):
-    return f'RegisterFuncMap["match-{column.lower()}"] = Filter{column.capitalize()}Match'
+    onerow = f"\t{c1}  string\n"
+    if c1 in repeated:
+        onerow = f"\t{c1}    uint16\n"
+    columnsItem.append(onerow)
 
 
-def create_register_contains_func(column):
-    return f'RegisterFuncMap["contains-{column.lower()}"] = Filter{column.capitalize()}Contains'
+# create Shrink code for repeated fields
+# where we map uint16 to a string value.
+shrinkVars = []
+shrinkItems = []
+shrinkvartemplate = env.get_template('shrinkVars.jinja2')
+shrinktemplate = env.get_template('shrinkColumn.jinja2')
+for c in repeated:
+    shrinkVars.append(
+        shrinkvartemplate.render(column=c, bitarray=c in bitarray))
+    shrinkItems.append(shrinktemplate.render(column=c))
 
 
-def create_register_startswith_func(column):
-    return f'RegisterFuncMap["startswith-{column.lower()}"] = Filter{column.capitalize()}StartsWith'
+# create the actual shrinked/expand Item fields.
+shrinkItemFields = []
+expandItemFields = []
+
+for c in allcolumns:
+
+    if c in ignored:
+        continue
+
+    if c in repeated:
+        # string to unint
+        shrinkItemFields.append(f"\t\t{c}IdxMap[i.{c}],\n")
+        # unint back to string
+        expandItemFields.append(f"\t\t{c}[i.{c}],\n")
+    else:
+        shrinkItemFields.append(f"\t\ti.{c},\n")
+        expandItemFields.append(f"\t\ti.{c},\n")
 
 
-def create_register_getter(column):
-    return f'RegisterGetters["{column.lower()}"] = Getters{column.capitalize()}'
+# ItemIn Columns
+inColumns = []
+for c in allcolumns_org:
+    inColumns.append(f'\t\t"{c}",\n')
+
+# ItemOut Columns
+outColumns = []
+for cc, c in zip(allcolumns, allcolumns_org):
+    # cc CamelCaseColumn.
+    if cc in ignored:
+        continue
+    outColumns.append(f'\t\t"{c}",\n')
+
+# create column filters.
+# match, startswith, contains etc
+
+columnFilters = []
+filtertemplate = env.get_template("filters.jinja2")
+
+for c in allcolumns:
+    if c in ignored:
+        continue
+
+    lookup = f"i.{c}"
+    if c in repeated:
+        lookup = f"{c}[i.{c}]"
+
+    txt = filtertemplate.render(column=c, lookup=lookup)
+    columnFilters.append(txt)
+
+registerFilters = []
+rtempl = env.get_template('registerFilters.jinja2')
+# register filters
+for c, co in zip(allcolumns, allcolumns_org):
+    if c in ignored:
+        continue
+    txt = rtempl.render(co=co, columnName=c, bitarray=c in bitarray)
+    registerFilters.append(txt)
+
+sortColumns = []
+sortTemplate = env.get_template('sortfunc.jinja2')
+
+# create sort functions
+for c, co in zip(allcolumns, allcolumns_org):
+    if c in ignored:
+        continue
+
+    c1 = f"items[i].{c} < items[j].{c}"
+    c2 = f"items[i].{c} > items[j].{c}"
+
+    if c in repeated:
+        c1 = f"{c}[items[i].{c}] < {c}[items[j].{c}]"
+        c2 = f"{c}[items[i].{c}] > {c}[items[j].{c}]"
+
+    txt = sortTemplate.render(co=co, c1=c1, c2=c2)
+    sortColumns.append(txt)
 
 
-def create_register_groupby(column):
-    return f'RegisterGroupBy["{column.lower()}"] = Getters{column.capitalize()}'
+csv_columns = []
+for c in allcolumns:
+    csv_columns.append(f'\t"{c}",\n')
 
 
-def create_register_reduce(column):
-    return 'RegisterReduce["count"] = reduceCount'
+# Finally render the model.go template
+modeltemplate = env.get_template('model.template.jinja2')
 
+geometryGetter = '""'
+print('GEOCOLUMNS: ' + " ".join(geocolumns))
+if len(geocolumns) == 1:
+    geometryGetter = f"Getters{geocolumns[0]}(&i)"
 
-def create_grouped():
-    return """
-type GroupedOperations struct {
-	Funcs   registerFuncType
-	GroupBy registerGroupByFunc
-	Getters registerGettersMap
-	Reduce  registerReduce
-}
+output = modeltemplate.render(
+    initRepeatColumns=''.join(initRepeatColumns),
+    columnsItemIn=''.join(columnsItemIn),
+    columnsItemOut=''.join(columnsItemOut),
+    columnsItem=''.join(columnsItem),
+    shrinkVars=''.join(shrinkVars),
+    shrinkItems=''.join(shrinkItems),
+    shrinkItemFields=''.join(shrinkItemFields),
+    expandItemFields=''.join(expandItemFields),
+    csv_columns=''.join(csv_columns),
+    inColumns=''.join(inColumns),
+    outColumns=''.join(outColumns),
+    columnFilters=''.join(columnFilters),
+    registerFilters=''.join(registerFilters),
+    sortColumns=''.join(sortColumns),
+    indexcolumn=allcolumns[index],
+    geometryGetter=geometryGetter,
+    bitArrayStores=''.join(bitArrayStores),
+    bitArrayGetters=''.join(bitArrayGetters),
+)
 
-var Operations GroupedOperations
+f = open('model.go', 'w')
+f.write(output)
+f.close()
 
-var RegisterFuncMap registerFuncType
-var RegisterGroupBy registerGroupByFunc
-var RegisterGetters registerGettersMap
-var RegisterReduce registerReduce
-"""
-
-def create_sortby_line_plus(column):
-	return f'"{column.lower()}"' + ": func(i, j int) bool { return " + f"items[i].{column.capitalize()} < items[j].{column.capitalize()} " + " },"
-
-def create_sortby_line_minus(column):
-	return f'"-{column.lower()}"' + ": func(i, j int) bool { return " + f"items[i].{column.capitalize()} > items[j].{column.capitalize()} " + " },"
-
-def create_sortby(row):
-    start = """func sortBy(items Items, sortingL []string) (Items, []string) {
-	sortFuncs := map[string]func(int, int) bool{"""
-    lines = []
-    for k in row.keys():
-        lines.append(create_sortby_line_plus(k))
-        lines.append(create_sortby_line_minus(k))
-        lines.append("\n")
-    lines.append("}")
-    end = """
-    for _, sortFuncName := range sortingL {
-    	sortFunc := sortFuncs[sortFuncName]
-        sort.Slice(items, sortFunc)
-                                }
-        // TODO must be nicer way
-        keys := []string{}
-        for key := range sortFuncs {
-              keys = append(keys, key)
-        }
-
-        return items, keys
-        }"""
-    return start + "\n".join(lines) + end
-
-if __name__ == "__main__":
-
-    filename = str(sys.argv[sys.argv.index('-f')+1]) if '-f' in sys.argv else "items.csv"
-    file_format = str(sys.argv[sys.argv.index('-format')+1]) if '-format' in sys.argv else "csv"
-
-    if file_format not in supported_fileformats():
-        print(f"{file_format} not part of supported file formats {','.join(supported_fileformats())}")
-        sys.exit()
-
-    with open(filename) as f:
-        reader = create_reader(f, file_format)
-        row = dict(next(reader))
-
-    print("package main")
-    print()
-
-    print("import (")
-    print('"sort"')
-    print('"strconv"')
-    print('"strings"')
-    print(")")
-    print(create_struct(row))
-    print()
-    print(create_columns(row))
-    print()
-    print(create_row(row))
-    print()
-    print(create_getindex(row))
-    print()
-
-    print("// contain filters")
-    for k in row.keys():
-        print(create_filter_contains(k))
-
-    print()
-    print("// startswith filters")
-    for k in row.keys():
-        print(create_filter_startswith(k))
-
-    print()
-    print("// match filters")
-    for k in row.keys():
-        print(create_filter_match(k))
-
-    print()
-    print("// reduce functions")
-    print(create_reduce(None))
-
-    print()
-    print("// getters")
-    for k in row.keys():
-        print(create_getter(k))
-    print()
-
-
-    print(create_grouped())
-    print("func init() {")
-    print(create_init_register())
-
-    print()
-    print("// register match filters")
-    for k in row.keys():
-        print(create_register_match_func(k))
-
-    print()
-    print("// register contains filters")
-    for k in row.keys():
-        print(create_register_contains_func(k))
-
-    print()
-    print("// register startswith filters")
-    for k in row.keys():
-        print(create_register_startswith_func(k))
-    print()
-
-    print()
-    print("// register getters ")
-    for k in row.keys():
-        print(create_register_getter(k))
-    print()
-
-    print()
-    print("// register groupby ")
-    for k in row.keys():
-        print(create_register_groupby(k))
-    print()
-
-
-    print()
-    print("// register reduce functions")
-    print(create_register_reduce(None))
-
-    print("}")
-
-    print(create_sortby(row))
-    print()
-
+print('saved in model.go')
+print('!!NOTE!! edit the default search filter')
